moved layer selection control into common
