+++ src/common/components/map/layer-selection-control/layer-selection-control.component.html
