+++ src/modules/growl/growl.component.html
@@ -37,9 +37,9 @@
       </ng-container>
       <ng-container controls>
         <mgl-control position="top-right">
-          <growl-layer-selection-control
+          <map-layer-selection-control
             [layers]="selectedLayers"
-          ></growl-layer-selection-control>
+          ></map-layer-selection-control>
         </mgl-control>
         <mgl-control
           mglNavigation
