+++ src/modules/growl/growl.component.ts
@@ -25,7 +25,7 @@ import {StyleSpecification} from "maplibre-gl";
 import * as turf from "@turf/turf";
 
 import {DisplayInfoControlComponent} from "./map/display-info-control/display-info-control.component";
-import {LayerSelectionControlComponent} from "./map/layer-selection-control/layer-selection-control.component";
+import {LayerSelectionControlComponent} from "../../common/components/map/layer-selection-control/layer-selection-control.component";
 import {LegendControlComponent} from "./map/legend-control/legend-control.component";
 import {GrowlService} from "./growl.service";
 import {WithdrawalInfoControlComponent} from "./map/withdrawal-info-control/withdrawal-info-control.component";
