+++ src/modules/oowv/action-map/action-map.component.html
@@ -8,6 +8,11 @@
       [canvasContextAttributes]="{preserveDrawingBuffer: true}"
       [attributionControl]="false"
     >
+    <mgl-control position="top-right">
+      <map-layer-selection-control
+        [layers]="{}"
+      ></map-layer-selection-control>
+    </mgl-control>
     </mgl-map>
   </div>
 </div>
